--- app/src/main/java/com/example/mobileterm/CalendarFragment.java
@@ -1,57 +1,0 @@
-package com.example.mobileterm;
-
-import android.annotation.SuppressLint;
-import android.content.Context;
-import android.graphics.Color;
-import android.graphics.drawable.Drawable;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.CalendarView;
-import android.widget.EditText;
-import android.widget.RadioButton;
-import android.widget.RadioGroup;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.content.ContextCompat;
-
-import com.prolificinteractive.materialcalendarview.CalendarDay;
-import com.prolificinteractive.materialcalendarview.DayViewDecorator;
-import com.prolificinteractive.materialcalendarview.DayViewFacade;
-import com.prolificinteractive.materialcalendarview.MaterialCalendarView;
-import com.prolificinteractive.materialcalendarview.OnDateSelectedListener;
-
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.util.Collection;
-import java.util.Collections;
-
-
-public class CalendarFragment extends Fragment {
-    iCalendarFragment iCalendarFragment;
-    gCalendarFragment gCalendarFragment;
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        iCalendarFragment = new iCalendarFragment();
-        getSupportFragmentManager().beginTransaction().replace(R.id.container,iCalendarFragment).commit();
-        gCalendarFragment = new gCalendarFragment();
-        return inflater.inflate(R.layout.fragment_calendar, container, false);
-    }
-
-    public void onFragmentChanged(int index) {
-        if (index == 0) {
-            getSupportFragmentManager().beginTransaction().replace(R.id.container, iCalendarFragment).commit();
-        } else if (index == 1) {
-            getSupportFragmentManager().beginTransaction().replace(R.id.container, gCalendarFragment).commit();
-        }
-    }
-}
