+++ app/src/main/java/com/example/mobileterm/Init/AuthActivity.java
@@ -1,0 +1,38 @@
+package com.example.mobileterm.Init;
+
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import com.example.mobileterm.R;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.firestore.FirebaseFirestore;
+
+public class AuthActivity extends AppCompatActivity {
+    private FirebaseAuth firebaseAuth;
+    private FirebaseUser user;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_auth);
+        ActionBar actionBar = getSupportActionBar();
+        actionBar.hide();
+
+        firebaseAuth = FirebaseAuth.getInstance();
+        user = firebaseAuth.getCurrentUser();
+        if (user != null){
+            FirebaseFirestore fv = FirebaseFirestore.getInstance();
+
+        }else{
+            StartActivity(LoginActivity.class);
+        }
+    }
+
+    private void StartActivity(Class c){
+        Intent intent = new Intent(this, c);
+        startActivity(intent);
+    }
+}
