+++ app/src/main/java/com/example/mobileterm/Init/LoginActivity.java
@@ -1,0 +1,15 @@
+package com.example.mobileterm.Init;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.mobileterm.R;
+import android.os.Bundle;
+
+public class LoginActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+    }
+}
