+++ app/src/main/res/layout/activity_auth.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".Init.AuthActivity"
+    android:background="@drawable/ic_launcher_foreground">
+
+</androidx.constraintlayout.widget.ConstraintLayout>
